# apps/api/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Jejak API",
    "version": "0.1.0",
    "description": "Contract-first business API for consent-bound settlement receivable financing."
  },
...
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local sandbox"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/v1/portfolio/summary": {
      "parameters": [
        {
          "$ref": "#/components/parameters/TenantId"
        }
      ],
      "get": {
        "operationId": "getPortfolioSummary",
        "summary": "Read tenant-scoped exposure and performance",
        "tags": [
          "Portfolio"
        ],
        "x-jejak-roles": [
          "FACILITY",
          "ORIGINATOR"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/CorrelationId"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Ok"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/v1/audit-events": {
      "parameters": [
        {
          "$ref": "#/components/parameters/TenantId"
        }
      ],
      "get": {
        "operationId": "listAuditEvents",
        "summary": "Search the append-only tenant audit log",
        "tags": [
          "Audit"
        ],
        "x-jejak-roles": [
          "ADMIN"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Cursor"
          },
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/CorrelationId"
          },
          {
            "name": "action",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9._-]{1,100}$"
            }
          },
          {
            "name": "resourceType",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^[A-Z0-9_]{1,64}$"
            }
          },
          {
            "name": "result",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "SUCCESS",
                "FAILURE"
              ]
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Ok"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Supabase access token. Authorization is resolved from backend memberships, not JWT user metadata."
      }
    },
    "schemas": {
...
      "uuidV7": {
        "type": "string",
        "format": "uuid-v7"
      },
      "timestamp.schema": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "$id": "https://jejak.finance/schemas/common/timestamp.schema.json",
        "title": "UtcTimestamp",
        "type": "string",
        "format": "utc-rfc3339"
      },
      "success-envelope.schema": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "$id": "https://jejak.finance/schemas/api/success-envelope.schema.json",
        "title": "ApiSuccess",
        "type": "object",
        "additionalProperties": false,
        "required": [
          "data",
          "meta"
        ],
        "properties": {
          "data": {},
          "meta": {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "requestId",
              "timestamp",
              "sandbox"
            ],
            "properties": {
              "requestId": {
                "$ref": "#/components/schemas/uuidV7"
              },
              "timestamp": {
                "$ref": "#/components/schemas/timestamp.schema"
              },
              "sandbox": {
                "type": "boolean"
              },
              "nextCursor": {
                "type": "string",
                "minLength": 1
              }
            }
          }
        }
      },
      "error-code.schema": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "$id": "https://jejak.finance/schemas/enums/error-code.schema.json",
        "title": "ErrorCode",
        "type": "string",
        "enum": [
          "AUTH_REQUIRED",
          "FORBIDDEN",
          "NOT_FOUND",
          "VALIDATION_FAILED",
          "VERSION_CONFLICT",
          "IDEMPOTENCY_CONFLICT",
          "INVALID_STATE_TRANSITION",
          "ATTESTATION_MISSING",
          "ATTESTATION_EXPIRED",
          "ATTESTATION_REVOKED",
          "CONTROL_NOT_VERIFIED",
          "CLAIM_ALREADY_ENCUMBERED",
          "INSUFFICIENT_FACILITY_LIQUIDITY",
          "HOLDER_NOT_AUTHORIZED",
          "ASSET_OPERATION_FAILED",
          "PARTNER_TIMEOUT",
          "PARTNER_REJECTED",
          "SETTLEMENT_DUPLICATE",
          "WATERFALL_INVARIANT_FAILED",
          "CIRCUIT_BREAKER_ACTIVE",
          "INVITATION_INVALID",
          "INVITATION_EXPIRED",
          "INVITATION_REVOKED",
          "INTERNAL_ERROR"
        ]
      },
      "error-envelope.schema": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "$id": "https://jejak.finance/schemas/api/error-envelope.schema.json",
        "title": "ApiError",
        "type": "object",
        "additionalProperties": false,
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "code",
              "message",
              "requestId",
              "retryable"
            ],
            "properties": {
              "code": {
                "$ref": "#/components/schemas/error-code.schema"
              },
              "message": {
                "type": "string",
                "minLength": 1
              },
              "requestId": {
                "$ref": "#/components/schemas/uuidV7"
              },
              "retryable": {
                "type": "boolean"
              },
              "details": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      },
...
    },
    "headers": {
      "RequestId": {
        "description": "UUIDv7 request identifier used for support and tracing.",
        "schema": {
          "type": "string",
          "format": "uuid-v7"
        }
      },
      "Sandbox": {
        "description": "Whether this response was produced using sandbox behavior.",
        "schema": {
          "type": "boolean"
        }
      }
    },
    "responses": {
      "Ok": {
        "description": "Successful operation.",
        "headers": {
          "X-Request-Id": {
            "$ref": "#/components/headers/RequestId"
          },
          "X-Jejak-Sandbox": {
            "$ref": "#/components/headers/Sandbox"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/success-envelope.schema"
            }
          }
        }
      },
      "BadRequest": {
        "description": "Invalid input.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/error-envelope.schema"
            }
          }
        }
      },
...
      "Unauthorized": {
        "description": "Authentication is required or invalid.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/error-envelope.schema"
            }
          }
        }
      },
      "Forbidden": {
        "description": "The authenticated actor lacks the required tenant, role, or object permission.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/error-envelope.schema"
            }
          }
        }
      },
...
    },
    "parameters": {
      "TenantId": {
        "name": "X-Jejak-Tenant-Id",
        "in": "header",
        "required": true,
        "description": "Explicit active tenant selected by the authenticated actor.",
        "schema": {
          "type": "string",
          "format": "uuid-v7"
        }
      },
...
      "CorrelationId": {
        "name": "X-Correlation-Id",
        "in": "header",
        "required": false,
        "schema": {
          "type": "string",
          "maxLength": 255
        }
      },
...
      "Cursor": {
        "name": "cursor",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string",
          "minLength": 1
        }
      },
      "Limit": {
        "name": "limit",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 50
        }
      },
...
    },
...
  }
}